backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
pymongo==4.6.1
motor==3.3.2
beanie==1.25.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.5.3
python-multipart==0.0.6

# Payment Gateway
razorpay==1.4.1

# Real-time (WebSocket)
websockets==12.0

# OTP & Notifications
twilio==8.13.0

# AWS S3 (Image uploads)
boto3==1.34.25

# Environment Variables
python-dotenv==1.0.0

# Validation & Utilities
pydantic-settings==2.1.0
email-validator==2.1.0
phonenumbers==8.13.28

# CORS
starlette==0.35.1

# Date/Time
python-dateutil==2.8.2

# Background Tasks
celery==5.3.6
redis==5.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
